+++ src/config/local.sample.js
@@ -2,8 +2,8 @@ export default {
   webUrl: '//localhost:3000',
 
   api: {
-    // endpoint: '//api.gw2armory.com/',
-    endpoint: '//localhost/',
+    endpoint: '//api.gw2armory.com/',
+    // endpoint: '//localhost/',
   },
 
   features: {